Phase 3: type-aware translation + 2936 new methods (509 files)
Translator improvements (ghidra_to_csharp.py Phase 3):
- Type-aware array offset guards: only emit .Length/[0] for actual arrays
- Narrowing cast insertion for non-literal assignments (int→byte/enum)
- get_Item/set_Item → C# indexer resolution
- Integrated --unity-check with iterative visibility auto-fixer
- Multi-level prefix capture fix for bool negation patterns

Results: 9681 methods translated, 0 Roslyn errors, 5266 Unity errors
(down from 6316 with 2936 additional methods written)

+++ Assets/Scripts/Dpr/Battle/Logic/BTL_ACTION.cs
@@ -14,15 +14,15 @@ namespace Dpr.Battle.Logic
 			if (!forceGWaza) {
 			  uVar2 = 0;
 			}
-			p = (ulong)(uint)(pokeID << 4) & 0x1f0 | (ulong)((waza & 0xffff) << 0xd) | uVar1 |
-			           (ulong)(uint)(targetPos << 9) & 0x1e00 | uVar2 | 1;
+			p = (ulong)(uint)(pokeID << 4) & 0x1f0 | (ulong)(((int)waza & 0xffff) << 0xd) | uVar1 |
+			           (ulong)(uint)((int)targetPos << 9) & 0x1e00 | uVar2 | 1;
 		}
 		
 		public static void ChangeFightTargetPos(ref BTL_ACTION_PARAM p, BtlPokePos nextTargetPos)
 		{
 			ulong uVar1 = default;
-			if (((nextTargetPos & 0xff) != 5) && (uVar1 = p, (uVar1 & 0xf) == 1)) {
-			  p = uVar1 & 0xffffffffffffe000 | uVar1 & 0x1ff | ((ulong)(nextTargetPos & 0x7fffff) & 0xf) << 9
+			if ((((int)nextTargetPos & 0xff) != 5) && (uVar1 = p, (uVar1 & 0xf) == 1)) {
+			  p = uVar1 & 0xffffffffffffe000 | uVar1 & 0x1ff | ((ulong)((int)nextTargetPos & 0x7fffff) & 0xf) << 9
 			  ;
 			}
 		}
